Converted the sample on a console app
