+++ Caching.sln
@@ -39,6 +39,8 @@ Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "SqlServerCacheConcurrencySa
 EndProject
 Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Microsoft.Extensions.Caching.SqlServer.Tests", "test\Microsoft.Extensions.Caching.SqlServer.Tests\Microsoft.Extensions.Caching.SqlServer.Tests.xproj", "{38EFEDCC-1076-4E1F-89C5-443C7FCCEE6A}"
 EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "MemoryCacheFileWatchSample", "samples\MemoryCacheFileWatchSample\MemoryCacheFileWatchSample.xproj", "{CD2A826A-A474-4491-94C8-43C557C68067}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -97,6 +99,10 @@ Global
 		{38EFEDCC-1076-4E1F-89C5-443C7FCCEE6A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{38EFEDCC-1076-4E1F-89C5-443C7FCCEE6A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{38EFEDCC-1076-4E1F-89C5-443C7FCCEE6A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CD2A826A-A474-4491-94C8-43C557C68067}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CD2A826A-A474-4491-94C8-43C557C68067}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CD2A826A-A474-4491-94C8-43C557C68067}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CD2A826A-A474-4491-94C8-43C557C68067}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -115,5 +121,6 @@ Global
 		{FED2ABB6-D0B3-4AC7-940F-250C9767EACB} = {459E1593-2C11-42CB-AD17-F7597E69E5D2}
 		{C94F2284-AA7A-4DC6-AC90-EAE0BF0DCFAB} = {459E1593-2C11-42CB-AD17-F7597E69E5D2}
 		{38EFEDCC-1076-4E1F-89C5-443C7FCCEE6A} = {2E14F082-1A04-4E7C-9162-1675119F9A9E}
+		{CD2A826A-A474-4491-94C8-43C557C68067} = {459E1593-2C11-42CB-AD17-F7597E69E5D2}
 	EndGlobalSection
 EndGlobal
