--- samples/MemoryCacheFileWatchSample/Abstractions/IGreetingService.cs
@@ -1,10 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-namespace MemoryCacheFileWatchSample.Abstractions
-{
-    public interface IGreetingService
-    {
-        string Greet(string recipient);
-    }
-}
