--- samples/MemoryCacheFileWatchSample/Abstractions/ITimeService.cs
@@ -1,12 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using System;
-
-namespace MemoryCacheFileWatchSample.Abstractions
-{
-    public interface ITimeService
-    {
-        DateTime Now { get; }
-    }
-}
