--- samples/MemoryCacheFileWatchSample/Services/TimeService.cs
@@ -1,20 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using System;
-using MemoryCacheFileWatchSample.Abstractions;
-
-
-namespace MemoryCacheFileWatchSample.Services
-{
-    public class TimeService : ITimeService
-    {
-        public DateTime Now
-        {
-            get
-            {
-                return DateTime.Now;
-            }
-        }
-    }
-}
