--- samples/MemoryCacheFileWatchSample/Startup.cs
@@ -1,37 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
-using MemoryCacheFileWatchSample.Abstractions;
-using MemoryCacheFileWatchSample.Services;
-using MemoryCacheFileWatchSample.Middleware;
-
-namespace MemoryCacheFileWatchSample
-{
-    public class Startup
-    {
-        // This method gets called by the runtime. Use this method to add services to the container.
-        // For more information on how to configure your application, visit http://go.microsoft.com/fwlink/?LinkID=398940
-        public void ConfigureServices(IServiceCollection services)
-        {
-            services.AddMemoryCache();
-            services.AddTransient<ITimeService, TimeService>();
-            services.AddTransient<IGreetingService, GreetingService>();
-        }
-
-        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env, ILoggerFactory loggerFactory)
-        {
-            loggerFactory.AddConsole();
-            app.UseGreetingMiddleware();
-
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-            }
-        }
-    }
-}
